+++ storybook/stories/staking/Rewards.stories.tsx
@@ -3,7 +3,7 @@ import { map } from 'lodash';
 import BigNumber from 'bignumber.js';
 import { action } from '@storybook/addon-actions';
 // Screens
-import StakingRewards from '../../../source/renderer/app/components/staking/rewards/StakingRewards';
+import { StakingRewards } from '../../../source/renderer/app/components/staking/rewards/StakingRewards';
 // Dummy data initialization
 import REWARDS from '../../../source/renderer/app/config/stakingRewards.dummy.json';